autopost/requirements.txt
tweepy>=4.14
asyncpraw>=7.7
feedparser>=6.0
httpx>=0.27
beautifulsoup4>=4.12
Pillow>=10.0
pyyaml>=6.0
python-dotenv>=1.0
loguru>=0.7
apscheduler>=3.10
